requirements.txt
## The following requirements were added by pip freeze:
boto==2.45.0
bz2file==0.98
decorator==4.0.11
elasticsearch==5.4.0
elasticsearch-dsl==5.1.0
gensim==0.13.4.1
isodate==0.5.4
joblib==0.10.3
langdetect==1.0.7
networkx==1.11
nltk==3.2.1
numpy==1.12.0
pandas==0.19.2
pkg-resources==0.0.0
pyaml==16.12.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
rdflib==4.2.1
requests==2.12.4
scipy==0.18.1
six==1.10.0
smart-open==1.3.5
urllib3==1.21.1
